+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.110.150](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.198.17...@standardnotes/desktop@3.110.150) (2025-09-04)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.149](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.198.16...@standardnotes/desktop@3.110.149) (2025-08-26)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.148](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.198.15...@standardnotes/desktop@3.110.148) (2025-08-25)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.110.147](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.198.14...@standardnotes/desktop@3.110.147) (2025-08-11)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.110.146](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.198.13...@standardnotes/desktop@3.110.146) (2025-08-05)
 
 ### Bug Fixes