+++ packages/services/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.71.10](https://github.com/standardnotes/app/compare/@standardnotes/services@1.71.9...@standardnotes/services@1.71.10) (2025-08-26)
+
+**Note:** Version bump only for package @standardnotes/services
+
 ## [1.71.9](https://github.com/standardnotes/app/compare/@standardnotes/services@1.71.8...@standardnotes/services@1.71.9) (2025-07-15)
 
 **Note:** Version bump only for package @standardnotes/services