+++ packages/ui-services/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.38.12](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.38.11...@standardnotes/ui-services@1.38.12) (2025-08-26)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
 ## [1.38.11](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.38.10...@standardnotes/ui-services@1.38.11) (2025-07-15)
 
 **Note:** Version bump only for package @standardnotes/ui-services